requirements.txt
faiss-cpu
langchain-huggingface
huggingface_hub
pypdf2
langchain-community
streamlit  
#pip install InstructorEmbedding sentence_transformers
asyncio
fitz
